user/api/requirements.txt
langchain==0.1.16
langchain-community==0.0.33
langchain-openai==0.1.3
langchain-pinecone==0.1.0
langchain-text-splitters==0.0.1
openai==1.20.0
pinecone-client==3.2.2
python-dotenv==1.0.1
pypdf==4.2.0